Data-Engineer-Projects/BigData-Ecommerce-Project/requirements.txt
# Core pipeline dependencies
apache-beam[gcp]==2.48.0
google-cloud-bigtable==2.19.0
google-cloud-bigquery>=3.10,<4
google-cloud-pubsub>=2.20,<3

# Production monitoring and operations
google-cloud-secret-manager>=2.16.0
google-cloud-monitoring>=2.18.0
google-cloud-logging>=3.5.0
google-cloud-trace>=1.10.0
google-cloud-billing>=1.4.0

# Data processing and analysis
PyYAML>=6.0
pandas>=1.5.0
matplotlib>=3.7.0

# Additional utilities
requests>=2.28.0
python-dateutil>=2.8.0